chore: apply feedbacks
