+++ grpc/src/main/java/com/linecorp/armeria/client/grpc/GrpcClientBuilder.java
@@ -74,7 +74,6 @@ import com.linecorp.armeria.common.grpc.GrpcJsonMarshallerBuilder;
 import com.linecorp.armeria.common.grpc.GrpcSerializationFormats;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageDeframer;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageFramer;
-import com.linecorp.armeria.internal.common.grpc.DefaultGrpcExceptionHandlerFunction;
 import com.linecorp.armeria.unsafe.grpc.GrpcUnsafeBufferUtil;
 
 import io.grpc.CallCredentials;
@@ -395,9 +394,9 @@ public final class GrpcClientBuilder extends AbstractClientOptionsBuilder {
             option(INTERCEPTORS.newValue(clientInterceptors));
         }
         if (exceptionHandler == null) {
-            exceptionHandler = DefaultGrpcExceptionHandlerFunction.ofDefault();
+            exceptionHandler = GrpcExceptionHandlerFunction.ofDefault();
         } else {
-            exceptionHandler = exceptionHandler.orElse(DefaultGrpcExceptionHandlerFunction.ofDefault());
+            exceptionHandler = exceptionHandler.orElse(GrpcExceptionHandlerFunction.ofDefault());
         }
         option(EXCEPTION_HANDLER.newValue(exceptionHandler));
