+++ grpc/src/main/java/com/linecorp/armeria/server/grpc/GrpcServiceBuilder.java
@@ -51,7 +51,6 @@ import com.linecorp.armeria.common.grpc.GrpcSerializationFormats;
 import com.linecorp.armeria.common.grpc.GrpcStatusFunction;
 import com.linecorp.armeria.common.grpc.protocol.AbstractMessageDeframer;
 import com.linecorp.armeria.common.grpc.protocol.ArmeriaMessageFramer;
-import com.linecorp.armeria.internal.common.grpc.DefaultGrpcExceptionHandlerFunction;
 import com.linecorp.armeria.server.HttpService;
 import com.linecorp.armeria.server.HttpServiceWithRoutes;
 import com.linecorp.armeria.server.Server;
@@ -976,9 +975,9 @@ public final class GrpcServiceBuilder {
         }
 
         if (exceptionHandler == null) {
-            exceptionHandler = DefaultGrpcExceptionHandlerFunction.ofDefault();
+            exceptionHandler = GrpcExceptionHandlerFunction.ofDefault();
         } else {
-            exceptionHandler = exceptionHandler.orElse(DefaultGrpcExceptionHandlerFunction.ofDefault());
+            exceptionHandler = exceptionHandler.orElse(GrpcExceptionHandlerFunction.ofDefault());
         }
         final GrpcExceptionHandlerFunction grpcExceptionHandler;
         if (exceptionMappingsBuilder != null) {
